+++ src/main/java/com/bajaj/SpringSecurityLatestApplication.java
@@ -1,0 +1,19 @@
+package com.bajaj;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+//@EnableJpaRepositories("com.bajaj.*")
+//@ComponentScan(basePackages = { "com.bajaj.*" })
+//@EntityScan("com.bajaj.*")
+@SpringBootApplication
+public class SpringSecurityLatestApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringSecurityLatestApplication.class, args);
+	}
+
+}
