+++ src/main/java/com/bajaj/repository/OrderRepository.java
@@ -1,0 +1,8 @@
+package com.bajaj.repository;
+
+import com.bajaj.entity.OrderEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderRepository extends JpaRepository<OrderEntity,Integer> {
+
+}
