+++ src/main/java/com/bajaj/util/ReferalCodeGenerator.java
@@ -1,0 +1,19 @@
+package com.bajaj.util;
+
+import java.security.SecureRandom;
+import java.util.Random;
+
+public class ReferalCodeGenerator {
+    public static String code(int codeLength){
+        char[] chars = "abcdefghijklmnopqrstuvwxyz1234567890".toCharArray();
+        StringBuilder sb = new StringBuilder();
+        Random random = new SecureRandom();
+        for (int i = 0; i < codeLength; i++) {
+            char c = chars[random.nextInt(chars.length)];
+            sb.append(c);
+        }
+        String output = sb.toString();
+        System.out.println(output);
+        return output ;
+    }
+}
