--- src/main/java/com/javatechie/SpringSecurityLatestApplication.java
@@ -1,13 +1,0 @@
-package com.javatechie;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringSecurityLatestApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringSecurityLatestApplication.class, args);
-	}
-
-}
