--- src/main/java/com/javatechie/config/UserInfoUserDetailsService.java
@@ -1,26 +1,0 @@
-package com.javatechie.config;
-
-import com.javatechie.entity.UserEntity;
-import com.javatechie.repository.UserInfoRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-
-import java.util.Optional;
-
-@Component
-public class UserInfoUserDetailsService implements UserDetailsService {
-
-    @Autowired
-    private UserInfoRepository repository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Optional<UserEntity> userInfo = repository.findByName(username);
-        return userInfo.map(UserInfoUserDetails::new)
-                .orElseThrow(() -> new UsernameNotFoundException("user not found " + username));
-
-    }
-}
