--- src/main/java/com/javatechie/dto/AuthRequest.java
@@ -1,24 +1,0 @@
-package com.javatechie.dto;
-
-
-public class AuthRequest {
-
-    private String username ;
-    private String password;
-    public String getUsername() {
-        return username;
-    }
-    public void setUsername(String username) {
-        this.username = username;
-    }
-    public String getPassword() {
-        return password;
-    }
-    public void setPassword(String password) {
-        this.password = password;
-    }
-    @Override
-    public String toString() {
-        return "AuthRequest [username=" + username + ", password=" + password + "]";
-    }
-}
