--- src/main/java/com/javatechie/entity/UserEntity.java
@@ -1,36 +1,0 @@
-package com.javatechie.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name="userinfo")
-public class UserEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="id")
-    private int id;
-    @Column(name = "name")
-    private String name;
-    @Column(name="email")
-    private String email;
-    @Column(name="password")
-    private String password;
-    @Column(name="roles")
-    private String roles;
-    @Column
-    private int points;//Need to add to database
-    @Column
-    private String referalCode;
-}
-
