--- src/main/java/com/javatechie/repository/OrderRepository.java
@@ -1,7 +1,0 @@
-package com.javatechie.repository;
-
-import com.javatechie.entity.OrderEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<OrderEntity,Integer> {
-}
