--- src/main/java/com/javatechie/repository/ProductRepository.java
@@ -1,11 +1,0 @@
-package com.javatechie.repository;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.javatechie.entity.ProductEntity;
-
-public interface ProductRepository extends JpaRepository<ProductEntity, Integer> {
-    List findById(int productId);
-}
