+++ src/test/java/com/bajaj/SpringSecurityLatestApplicationTests.java
@@ -1,0 +1,13 @@
+package com.bajaj;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringSecurityLatestApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
