--- src/test/java/com/javatechie/SpringSecurityLatestApplicationTests.java
@@ -1,13 +1,0 @@
-package com.javatechie;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringSecurityLatestApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
